lift state to MovieListView
