+++ src/components/Footer.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Footer = () => {
+  return (
+    <footer>
+      <h2>Footer</h2>
+    </footer>
+  )
+}
+
+export default Footer
