+++ src/components/Header.js
@@ -1,0 +1,12 @@
+import React from 'react'
+
+const Header = () => {
+  return (
+    <header>
+      <h1>Marvin's Movies</h1>
+      <h2>Header</h2>
+    </header>
+  )
+}
+
+export default Header
