+++ src/components/MovieCard.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const MovieCard = () => {
+  return null
+}
+
+export default MovieCard
