+++ src/components/index.js
@@ -1,0 +1,4 @@
+export { default as Header } from './Header'
+export { default as MovieListView } from './MovieListView'
+export { default as Footer } from './Footer'
+export { default as MovieCard } from './MovieCard'
